requirements.txt
geopandas
rasterio
shapely
psycopg2-binary
requests
python-dotenv
supabase